requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
thetadata>=1.0.0
pandas>=1.5.0
numpy>=1.20.0